Format old codes
