+++ itchat/config.py
@@ -1,8 +1,8 @@
 import os, platform
 
-DIR = os.getcwd()
 BASE_URL = 'https://login.weixin.qq.com'
 OS = platform.system() #Windows, Linux, Darwin
+DIR = os.getcwd()
+DEFAULT_QR = 'QR.jpg'
 
-WELCOME_WORDS = 'Hello World!'
-USER_AGENT = u'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/54.0.2840.71 Safari/537.36'
+USER_AGENT = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_11_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/54.0.2840.71 Safari/537.36'
